+++ packages/transaction-builders/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @unipasswallet/transaction-builders
 
+## 0.0.1-alpha.22
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.27
+  - @unipasswallet/keys@0.0.1-alpha.27
+  - @unipasswallet/transactions@0.0.1-alpha.26
+  - @unipasswallet/utils@0.0.1-alpha.27
+  - @unipasswallet/wallet@0.0.1-alpha.27
+
+## 0.0.1-alpha.21
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.26
+  - @unipasswallet/keys@0.0.1-alpha.26
+  - @unipasswallet/transactions@0.0.1-alpha.25
+  - @unipasswallet/utils@0.0.1-alpha.26
+  - @unipasswallet/wallet@0.0.1-alpha.26
+
 ## 0.0.1-alpha.20
 
 ### Patch Changes