+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @unipasswallet/utils
 
+## 0.0.1-alpha.27
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.27
+
+## 0.0.1-alpha.26
+
+### Patch Changes
+
+- update versions
+- Updated dependencies
+  - @unipasswallet/abi@0.0.1-alpha.26
+
 ## 0.0.1-alpha.25
 
 ### Patch Changes